+++ authlete/dto/introspection_response.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@ class IntrospectionResponse(ApiResponse):
             'action':                IntrospectionAction,
             'clientId':              int,
             'subject':               str,
-            'scopes':                str,  # list of str
+            'scopes':                str,       # list of str
             'existent':              bool,
             'usable':                bool,
             'sufficient':            bool,
@@ -36,7 +36,9 @@ class IntrospectionResponse(ApiResponse):
             'properties':            Property,  # list of Property
             'clientIdAlias':         str,
             'clientIdAliasUsed':     bool,
-            'certificateThumbprint': str
+            'certificateThumbprint': str,
+            'resources':             str,       # list of str
+            'accessTokenResources':  str,       # list of str
         }
 
         super().__init__(nameAndValues, nameAndTypes)